+++ clang-tools-extra/test/clang-tidy/bugprone-argument-comment-gmock.cpp
@@ -1,4 +1,4 @@
-// RUN: %check_clang_tidy %s misc-argument-comment %t
+// RUN: %check_clang_tidy %s bugprone-argument-comment %t
 
 namespace testing {
 namespace internal {